Changed database structure. Fixed login and registration view. Added validation on backend, added projects part.

+++ ZDoneWebApi/ZDoneWebApi/Controllers/ListController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
+using System.Linq;
 using System.Threading.Tasks;
 using ZDoneWebApi.BusinessLogic.Interfaces;
 using ZDoneWebApi.Data.DTOs;
@@ -16,17 +17,38 @@ namespace ZDoneWebApi.Controllers
     public class ListController : ControllerBase
     {
         private readonly IListBl _listBl;
+        private readonly IUserBl _userBl;
+        private string userId;
+        private int projectId;
 
-        public ListController(IListBl listBl)
+        public ListController(IListBl listBl, IUserBl userBl)
         {
             _listBl = listBl;
+            _userBl = userBl;
+        }
+
+        private (string, int) GetUserId()
+        {
+            var userId = User != null ? User.Claims.Single(c => c.Type == "id").Value : null;
+            var projectId = User != null ? User.Claims.Single(c => c.Type == "projectId").Value : null;
+            if (projectId != null)
+            {
+                return (userId, int.Parse(projectId));
+            }
+            return (userId, 0);
         }
 
         // GET: api/Item
         [HttpGet]
         public async Task<ActionResult> Get()
         {
+            (userId, projectId) = GetUserId();
+            if (!(await _userBl.IsHaveProjectPermission(userId, projectId)))
+            {
+                return Forbid();
+            }
             var allItems = await _listBl.GetAllAsync();
+
             if (allItems == null)
                 return NotFound();
             return Ok(allItems);
@@ -35,7 +57,12 @@ namespace ZDoneWebApi.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult> Get(int id)
         {
-            var list = await _listBl.ReadAsync(id);
+            (userId, projectId) = GetUserId();
+            if (!(await _userBl.IsHaveAccessToList(id, userId)))
+            {
+                return Forbid();
+            }
+            var list = await _listBl.ReadAsync(id, userId);
             if (list == null)
                 return NotFound();
             return Ok(list);
@@ -44,6 +71,11 @@ namespace ZDoneWebApi.Controllers
         [HttpGet("{id}/items")]
         public async Task<ActionResult> GetItems(int id)
         {
+            (userId, projectId) = GetUserId();
+            if (!(await _userBl.IsHaveAccessToList(id, userId)))
+            {
+                return Forbid();
+            }
             var list = await _listBl.GetItemsByListId(id);
             if (list == null)
                 return NotFound();
@@ -53,6 +85,11 @@ namespace ZDoneWebApi.Controllers
         [HttpGet("{id}/items/done")]
         public async Task<ActionResult> GetDoneItems(int id)
         {
+            (userId, projectId) = GetUserId();
+            if (!(await _userBl.IsHaveAccessToList(id, userId)))
+            {
+                return Forbid();
+            }
             var list = await _listBl.GetDoneItemsByListId(id);
             if (list == null)
                 return NotFound();
@@ -62,6 +99,11 @@ namespace ZDoneWebApi.Controllers
         [HttpGet("{id}/items/undone")]
         public async Task<ActionResult> GetUnDoneItems(int id)
         {
+            (userId, projectId) = GetUserId();
+            if (!(await _userBl.IsHaveAccessToList(id, userId)))
+            {
+                return Forbid();
+            }
             var list = await _listBl.GetUndoneItemsByListId(id);
             if (list == null)
                 return NotFound();
@@ -74,6 +116,11 @@ namespace ZDoneWebApi.Controllers
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+            (userId, projectId) = GetUserId();
+            if (!(await _userBl.IsHaveProjectPermission(userId, projectId)))
+            {
+                return Forbid();
+            }
             var result = await _listBl.CreateAsync(list);
             if (!result.Success)
                 return BadRequest(result.Message);
@@ -86,6 +133,12 @@ namespace ZDoneWebApi.Controllers
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
+            (userId, projectId) = GetUserId();
+            if (!(await _userBl.IsHaveProjectPermission(userId, projectId)))
+            {
+                return Forbid();
+            }
             var result = await _listBl.UpdateAsync(list);
             if (!result.Success)
                 return BadRequest(result.Message);
@@ -96,6 +149,11 @@ namespace ZDoneWebApi.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult> Delete(int id)
         {
+            (userId, projectId) = GetUserId();
+            if (!(await _userBl.IsHaveAccessToList(id, userId)))
+            {
+                return Forbid();
+            }
             var result = await _listBl.DeleteAsync(id);
             if (!result.Success)
                 return BadRequest(result.Message);
